fix: Change PostgreSQL timestamp column to DOUBLE PRECISION
- Slack timestamps are decimals (e.g., 1768240225.567979)
- PostgreSQL BIGINT doesn't accept decimals
- Changed to DOUBLE PRECISION to handle decimal timestamps
- This fixes the fallback to SQLite issue

+++ src/database/ChatHistory.ts
@@ -319,7 +319,7 @@ class PostgreSQLChatHistory extends ChatHistoryBackend {
           message_id VARCHAR(255) NOT NULL,
           thread_id VARCHAR(255),
           message_text TEXT NOT NULL,
-          timestamp BIGINT NOT NULL,
+          timestamp DOUBLE PRECISION NOT NULL,
           created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
           UNIQUE(platform, message_id)
         );
